refactor: consitent structure between index and table
Includes a bit of a behaviour change, no exception on page type. But we
depend on lots of other things being "right", so I think a bit odd to
raise an exception on these. There isn't some other page type that isn't
supported by the code, which would be a reasonable reason to raise an
exception.